feat: add ReplaceZeroWidthChars function
